fixed stack overflow issue
For some reason, the sizes of the panels don't always add up to the size of the panelGroup. This causes issues where the resizePanel method is called recursively and never ends leading to a stack overflow exception. The solution is to add the unacounted for space to the given panel.